+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.1] - 2019-02-28
 ### Added
 - World generation (flowers, netherwood, and crimson iron)
+### Changed
+- Crafting station part slots removed. This leaves just the crafting grid and side storage.
 ### Fixed
+- Blocks with GUIs can be opened again
 - Flower and netherwood sapling models corrected
 
 ## [1.0.0] - 2019-02-24